backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
neo4j==5.14.1
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.9
pdfplumber==0.10.3
httpx==0.25.2
python-multipart==0.0.6
sqlalchemy>=2.0.0
passlib>=1.7.4
bcrypt<4.0.0
pyjwt>=2.1.0
